Done Discount

+++ backend/src/main/java/com/threed_model_market/project/repository/DiscountRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 @Repository
 @SuppressWarnings("unused")
-public interface DiscountRepository extends JpaRepository<Discount, Integer> {
+public interface DiscountRepository extends JpaRepository<Discount, Long> {
     Optional<Discount> findByCode(String code);
 }
